requirements.txt
fastapi
uvicorn
uvicorn[standard]
python-dotenv
